create PostContent component
