+++ ui/src/components/Post/PostContent.js
@@ -1,0 +1,72 @@
+/*
+2023-03-15
+ui/src/components/Post/PostContent.js
+
+*/
+
+import React from 'react'
+import { styled } from '@mui/material/styles';
+import CardActions from '@mui/material/CardActions';
+import CardContent from '@mui/material/CardContent';
+import Collapse from '@mui/material/Collapse';
+import Typography from '@mui/material/Typography';
+import FavoriteIcon from '@mui/icons-material/Favorite';
+import IconButton from '@mui/material/IconButton';
+import ShareIcon from '@mui/icons-material/Share';
+import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
+
+/*
+This code is modified from a documentation guide on Material UI Card components from Material UI SAS 2023, retrieved 2023-03-13 from mui.com
+guide here
+https://mui.com/material-ui/react-card/#complex-interaction
+*/
+const ExpandMore = styled((props) => {
+    const { expand, ...other } = props;
+    return <IconButton {...other} />;
+    })(({ theme, expand }) => ({
+        transform: !expand ? 'rotate(0deg)' : 'rotate(180deg)',
+        marginLeft: 'auto',
+        transition: theme.transitions.create('transform', {
+            duration: theme.transitions.duration.shortest,
+    }),
+}));
+
+const PostContent = ({ description, content }) => {
+    const [expanded, setExpanded] = React.useState(false);
+
+    const handleExpandClick = () => {
+        setExpanded(!expanded);
+    };
+    return (<>
+    <CardContent>
+        <Typography variant="body2" color="text.secondary">
+            {description}
+        </Typography>
+    </CardContent>
+    <CardActions disableSpacing>
+        <IconButton aria-label="add to favorites">
+        <FavoriteIcon />
+        </IconButton>
+        <IconButton aria-label="share">
+        <ShareIcon />
+        </IconButton>
+        <ExpandMore
+        expand={expanded}
+        onClick={handleExpandClick}
+        aria-expanded={expanded}
+        aria-label="show more"
+        >
+        <ExpandMoreIcon />
+        </ExpandMore>
+    </CardActions>
+    <Collapse in={expanded} timeout="auto" unmountOnExit>
+        <CardContent>
+        <Typography paragraph>
+            {content}
+        </Typography>
+        </CardContent>
+    </Collapse>
+    </>);
+}
+
+export default PostContent
